refactor: 컨트롤러 네이밍 변경, 메서드 이동
