+++ src/main/java/roomescape/controller/AdminController.java
@@ -1,30 +1,48 @@
 package roomescape.controller;
 
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.Reservation;
+import roomescape.dto.AdminReservationSaveRequest;
+import roomescape.dto.MemberResponse;
+import roomescape.dto.ReservationResponse;
+import roomescape.dto.ReservationTimeResponse;
+import roomescape.dto.ThemeResponse;
+import roomescape.service.MemberService;
+import roomescape.service.ReservationService;
+import roomescape.service.ReservationTimeService;
+import roomescape.service.ThemeService;
 
-@RequestMapping("/admin")
-@Controller
+@RestController
 public class AdminController {
 
-    @GetMapping
-    public String mainPage() {
-        return "/admin/index";
+    private final MemberService memberService;
+    private final ReservationService reservationService;
+    private final ReservationTimeService reservationTimeService;
+    private final ThemeService themeService;
+
+    public AdminController(
+            final MemberService memberService,
+            final ReservationService reservationService,
+            final ReservationTimeService reservationTimeService,
+            final ThemeService themeService)
+    {
+        this.memberService = memberService;
+        this.reservationService = reservationService;
+        this.reservationTimeService = reservationTimeService;
+        this.themeService = themeService;
     }
 
-    @GetMapping("/reservation")
-    public String reservationPage() {
-        return "/admin/reservation-new";
-    }
+    @PostMapping("/admin/reservations")
+    public ResponseEntity<ReservationResponse> createReservation(@RequestBody final AdminReservationSaveRequest request) {
+        final MemberResponse memberResponse = memberService.findById(request.memberId());
+        final ReservationTimeResponse reservationTimeResponse = reservationTimeService.findById(request.timeId());
+        final ThemeResponse themeResponse = themeService.findById(request.themeId());
 
-    @GetMapping("/time")
-    public String timePage() {
-        return "/admin/time";
-    }
-
-    @GetMapping("/theme")
-    public String themePage() {
-        return "/admin/theme";
+        final Reservation reservation = request.toModel(memberResponse, themeResponse, reservationTimeResponse);
+        return ResponseEntity.status(HttpStatus.CREATED).body(reservationService.create(reservation));
     }
 }
