+++ src/main/java/roomescape/controller/PageController.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class UserController {
+public class PageController {
 
     @GetMapping("/")
     public String mainPage() {
