+++ src/main/java/roomescape/controller/ReservationController.java
@@ -2,15 +2,9 @@ package roomescape.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import roomescape.config.AuthenticationPrincipal;
 import roomescape.domain.Reservation;
-import roomescape.dto.AdminReservationSaveRequest;
 import roomescape.dto.MemberLoginResponse;
 import roomescape.dto.MemberReservationSaveRequest;
 import roomescape.dto.MemberResponse;
@@ -24,6 +18,7 @@ import roomescape.service.ThemeService;
 
 import java.util.List;
 
+@RequestMapping("/reservations")
 @RestController
 public class ReservationController {
 
@@ -44,7 +39,7 @@ public class ReservationController {
         this.themeService = themeService;
     }
 
-    @PostMapping("/reservations")
+    @PostMapping
     public ResponseEntity<ReservationResponse> createReservation(
             @AuthenticationPrincipal MemberLoginResponse loginResponse,
             @RequestBody final MemberReservationSaveRequest request) {
@@ -56,22 +51,12 @@ public class ReservationController {
         return ResponseEntity.status(HttpStatus.CREATED).body(reservationService.create(reservation));
     }
 
-    @PostMapping("/admin/reservations")
-    public ResponseEntity<ReservationResponse> createReservation(@RequestBody final AdminReservationSaveRequest request) {
-        final MemberResponse memberResponse = memberService.findById(request.memberId());
-        final ReservationTimeResponse reservationTimeResponse = reservationTimeService.findById(request.timeId());
-        final ThemeResponse themeResponse = themeService.findById(request.themeId());
-
-        final Reservation reservation = request.toModel(memberResponse, themeResponse, reservationTimeResponse);
-        return ResponseEntity.status(HttpStatus.CREATED).body(reservationService.create(reservation));
-    }
-
-    @GetMapping("/reservations")
+    @GetMapping
     public ResponseEntity<List<ReservationResponse>> findReservations() {
         return ResponseEntity.ok(reservationService.findAll());
     }
 
-    @DeleteMapping("/reservations/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable final Long id) {
         reservationService.delete(id);
         return ResponseEntity.noContent().build();
